requirement.txt.txt
streamlit
pandas
numpy
matplotlib
requests
openpyxl
python-dotenv
